Move PropertyDictionaryExtensions to Azure.Common
